+++ src/test/unit/java/org/sukrupa/bigneeds/BigNeedTest.java
@@ -12,7 +12,7 @@ public class BigNeedTest {
     public void shouldCreatePowerGeneratorBigNeed(){
         BigNeed powerGeneratorBigNeed = new BigNeed("Power Generator", 50000,1);
         assertThat(powerGeneratorBigNeed.getItemName(),is("Power Generator"));
-        assertThat(powerGeneratorBigNeed.getCost(), is(50000));
+        assertThat(powerGeneratorBigNeed.getCost(), is((double)50000));
          assertThat(powerGeneratorBigNeed.getPriority(),is(1));
         
     }
@@ -21,6 +21,6 @@ public class BigNeedTest {
     public void shouldCreateComputerBigNeed(){
         BigNeed computerBigNeed = new BigNeed("Computer", 120000,2);
         assertThat(computerBigNeed.getItemName(),is("Computer"));
-        assertThat(computerBigNeed.getCost(), is(120000));
+        assertThat(computerBigNeed.getCost(), is((double)120000));
     }
 }
